move proc-macro inside
