+++ static-metric/Cargo.toml
@@ -17,10 +17,7 @@ include = [
 proc-macro = true
 
 [dependencies]
-syn = { version = "1.0", features = ["full", "extra-traits"] }
-proc-macro2 = "1.0"
-quote = "1.0"
-lazy_static = "1.0"
+static-metric-proc-macros = { path = "./proc-macros" }
 
 [dev-dependencies]
 prometheus = { path = "../" }
