+++ static-metric/proc-macros/Cargo.toml
@@ -1,0 +1,23 @@
+[package]
+name = "static-metric-proc-macros"
+version = "0.1.0"
+authors = ["Yilin Chen <sticnarf@gmail.com>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[lib]
+proc-macro = true
+
+[dependencies]
+syn = { version = "1.0", features = ["full", "extra-traits"] }
+proc-macro2 = "1.0"
+quote = "1.0"
+lazy_static = "1.3"
+
+[dev-dependencies]
+prometheus = { path = "../../" }
+coarsetime = "0.1"
+
+[features]
+default = []
