+++ static-metric/proc-macros/src/auto_flush_from.rs
