+++ static-metric/proc-macros/src/register_macro.rs
