+++ static-metric/src/lib.rs
@@ -22,80 +22,6 @@ assert_eq!(HIGH_FIVE_COUNTER.get(), 1);
 Is it reccomended that you consult the [`prometheus` documentation for more information.](https://docs.rs/prometheus/)
 */
 
-extern crate lazy_static;
-extern crate proc_macro;
-extern crate proc_macro2;
-#[macro_use]
-extern crate quote;
-extern crate syn;
+extern crate static_metric_proc_macros;
 
-mod auto_flush_builder;
-mod auto_flush_from;
-mod builder;
-mod parser;
-mod register_macro;
-mod util;
-
-use proc_macro::TokenStream;
-
-use self::builder::TokensBuilder;
-use self::parser::StaticMetricMacroBody;
-use self::register_macro::RegisterMethodInvoking;
-use auto_flush_builder::AutoFlushTokensBuilder;
-use auto_flush_from::AutoFlushFromDef;
-
-/// Build static metrics.
-#[proc_macro]
-pub fn make_static_metric(input: TokenStream) -> TokenStream {
-    let body: StaticMetricMacroBody = syn::parse(input).unwrap();
-    TokensBuilder::build(body).into()
-}
-
-/// Build auto flush able static metrics.
-#[proc_macro]
-pub fn make_auto_flush_static_metric(input: TokenStream) -> TokenStream {
-    let body: StaticMetricMacroBody = syn::parse(input).unwrap();
-    AutoFlushTokensBuilder::build(body).into()
-}
-
-#[proc_macro]
-pub fn auto_flush_from(input: TokenStream) -> TokenStream {
-    let def: AutoFlushFromDef = syn::parse(input).unwrap();
-    def.auto_flush_from().into()
-}
-
-/// Register a `CounterVec` and create static metrics from it.
-#[proc_macro]
-pub fn register_static_counter_vec(input: TokenStream) -> TokenStream {
-    register_static_vec("counter", input)
-}
-
-/// Register a `IntCounterVec` and create static metrics from it.
-#[proc_macro]
-pub fn register_static_int_counter_vec(input: TokenStream) -> TokenStream {
-    register_static_vec("int_counter", input)
-}
-
-/// Register a `GaugeVec` and create static metrics from it.
-#[proc_macro]
-pub fn register_static_gauge_vec(input: TokenStream) -> TokenStream {
-    register_static_vec("gauge", input)
-}
-
-/// Register a `IntGaugeVec` and create static metrics from it.
-#[proc_macro]
-pub fn register_static_int_gauge_vec(input: TokenStream) -> TokenStream {
-    register_static_vec("int_gauge", input)
-}
-
-/// Register a `HistogramVec` and create static metrics from it.
-#[proc_macro]
-pub fn register_static_histogram_vec(input: TokenStream) -> TokenStream {
-    register_static_vec("histogram", input)
-}
-
-/// Procedural macro handler for `register_static_xxx_vec!`.
-fn register_static_vec(register_type: &str, input: TokenStream) -> TokenStream {
-    let invoking: RegisterMethodInvoking = syn::parse(input).unwrap();
-    invoking.into_tokens(register_type).into()
-}
+pub use static_metric_proc_macros::*;
